move core components to `Core\` namespace
