+++ app/Controller/AboutUs.php
@@ -2,6 +2,8 @@
 
 namespace App\Controller;
 
+use Core\AbstractController;
+
 class AboutUs extends AbstractController
 {
     public static function page(): void
