--- app/Route/Redirect.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace App\Route;
-
-class Redirect extends  Route
-{
-
-}
