--- app/Route/Route.php
@@ -1,22 +1,0 @@
-<?php
-
-namespace App\Route;
-
-class Route
-{
-    public static $validRoute = array();
-
-    public static function set($route, $method)
-    {
-        self::$validRoute = $route;
-
-        if (static::requestedUrl() == $route) {
-            $method->__invoke();
-        }
-    }
-
-    protected static function requestedUrl(): string
-    {
-        return parse_url($_SERVER['REQUEST_URI'], PHP_URL_PATH);
-    }
-}
