+++ src/Routing/Redirect.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Core\Routing;
+
+class Redirect extends Route
+{
+}
